refactor: no need to use Function separately from cron
