+++ sst/packages/functions/src/cron-jobs/redaction-digest/cron.ts
@@ -1,13 +1,74 @@
 import { Config } from 'sst/node/config'
 
-import { lambdaInvokerWithCronitor } from '../lambda-invoker'
+import { getSequelize } from '@/core/database/client'
+import { sendEmail } from '@/core/util/email'
 
-export async function handler() {
-  const sendRedactionDigest = lambdaInvokerWithCronitor(
-    Config.SEND_REDACTION_DIGEST_FUNCTION_URL,
-    Config.SEND_REDACTION_DIGEST_FUNCTION_NAME,
-    Config.SEND_REDACTION_DIGEST_FUNCTION_VERSION,
-    Config.CRONITOR_CODE_REDACTION_DIGEST,
+import { IS_LOCAL, LOCAL_DB_URI } from '../../env'
+
+import { createEmailBody, getUserRedactedCampaigns } from './helper'
+
+interface Event {
+  selectedRecipients?: Array<string>
+}
+export async function handler(event: Event) {
+  try {
+    if (IS_LOCAL) {
+      console.log('Running cron locally')
+      await sendRedactionDigest(event)
+      return
+    }
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const cronitor = require('cronitor')(Config.CRONITOR_URL_SUFFIX) // common to all jobs on our shared Cronitor account
+    const invokeFunction = cronitor.wrap(
+      Config.CRONITOR_CODE_REDACTION_DIGEST,
+      async function () {
+        await sendRedactionDigest(event)
+      },
+    )
+    await invokeFunction()
+  } catch (error) {
+    console.log(error)
+    throw error
+  }
+}
+
+export async function sendRedactionDigest(event: Event) {
+  // this is an optional parameter that allow for manually triggering the function for a specific set of users
+  const { selectedRecipients } = event
+  const dbUri = IS_LOCAL ? LOCAL_DB_URI : Config.POSTMAN_DB_URI
+  const sequelize = getSequelize(dbUri)
+  const userRedactedCampaigns = await getUserRedactedCampaigns(
+    sequelize,
+    selectedRecipients,
   )
-  await sendRedactionDigest()
+  const failedRecipients = []
+  for (const userCampaigns of userRedactedCampaigns) {
+    const { email, campaigns } = userCampaigns
+
+    try {
+      const emailBody = createEmailBody(campaigns)
+      await sendEmail({
+        recipient: email,
+        subject:
+          'Postman.gov.sg: Reminder for expiring campaign delivery reports',
+        body: emailBody,
+        tag: 'redaction-digest',
+      })
+
+      console.log(`Redaction reminder sent to ${email}`)
+    } catch (err) {
+      console.log(
+        `Failed to send redaction reminder to ${email}. Error: ${
+          (err as Error).message
+        }`,
+      )
+      failedRecipients.push(email)
+    }
+  }
+
+  if (failedRecipients.length > 0) {
+    throw new Error(
+      `Failed to send redaction reminders to ${failedRecipients.join(', ')}`,
+    )
+  }
 }
