+++ sst/packages/functions/src/cron-jobs/unsub-digest/cron.ts
@@ -1,13 +1,41 @@
 import { Config } from 'sst/node/config'
 
-import { lambdaInvokerWithCronitor } from '../lambda-invoker'
+import { getSequelize } from '@/core/database/client'
+
+import { IS_LOCAL, LOCAL_DB_URI } from '../../env'
+
+import { getUnsubscribeList, sendEmailAndUpdate } from './helper'
 
 export async function handler() {
-  const sendUnsubDigest = lambdaInvokerWithCronitor(
-    Config.SEND_UNSUB_DIGEST_FUNCTION_URL,
-    Config.SEND_UNSUB_DIGEST_FUNCTION_NAME,
-    Config.SEND_UNSUB_DIGEST_FUNCTION_VERSION,
-    Config.CRONITOR_CODE_UNSUB_DIGEST,
-  )
-  await sendUnsubDigest()
+  try {
+    if (IS_LOCAL) {
+      console.log('Running cron locally')
+      await sendUnsubDigest()
+      return
+    }
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const cronitor = require('cronitor')(Config.CRONITOR_URL_SUFFIX) // common to all jobs on our shared Cronitor account
+    const invokeFunction = cronitor.wrap(
+      Config.CRONITOR_CODE_UNSUB_DIGEST,
+      async function () {
+        await sendUnsubDigest()
+      },
+    )
+    await invokeFunction()
+  } catch (error) {
+    console.log(error)
+    throw error
+  }
+}
+
+async function sendUnsubDigest() {
+  const dbUri = IS_LOCAL ? LOCAL_DB_URI : Config.POSTMAN_DB_URI
+  const sequelize = getSequelize(dbUri)
+  // retrieve unsubscribed recipients grouped by campaigns and users
+  const unsubscribeDigests = await getUnsubscribeList(sequelize)
+
+  // generate email digest and send email to each user
+  for (const userDigest of unsubscribeDigests) {
+    await sendEmailAndUpdate(userDigest, sequelize)
+  }
 }
